verifier_service/requirements.txt

Django==5.1.5
djangorestframework==3.15.2
pdfplumber==0.11.4
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
